added: automatically add licenses to scripts
Thanks goes to Howard Baxton for his request to add the licenses to
the nPose scripts, which inspired the auto-add feature for licenses.

Each script of a project can have its own license added, see
bin/addlicense.pl.

To have a license added, create a file "license" in the project
directory.  For individual scripts, add a file "license-<script-name>"
(without extension).  The contents of the license file will be
prepended to the scripts (*.i and *.o).  The sources are not touched.