adjustment for license adding:
The script name doesn´t need to be in a comment in the first line of
the script anymore for automatic replacement to work.  It can now be
in any line.

It is simpler to change this than it is to make it so that the comment
with the script name remains in the first line when a license is
prepended.  The license should be at the top of the file anyway.