lsl-mode extended with functions to look up things on the wiki
Move point over a function name, and eww is used to display the
corresponding wiki page in another window --- bound to C-c h by
default.  A different wiki can be referred to by changing the key
binding, or by adding another one, or running 'lsl-lookup-lsl-ref
instead of 'lsl-lookup-lsl-ref2.

These functions are very much based on the GPL version of xlsl-mode.

I put this in because I started experimenting with auto-complete-mode
and yasnippet, and it doesn´t seem to make much sense to have those
without some sort of integrated documentation: You can easily bring in
a template for the "changed" event with yasnippet, but who remembers
all the relevant flags?  So now just do C-c h on it and you get the
wiki page for it.